+++ packages/docs/src/stories/Button.stories.tsx
@@ -90,6 +90,16 @@ export const Link: StoryObj<IButtonProps> = {
     controls: { exclude: ['Icon', 'iconSide', 'onlyIcon'] },
   },
 }
+
+export const Orange: StoryObj<IButtonProps> = {
+  args: {
+    variant: 'orange',
+  },
+  parameters: {
+    controls: { exclude: ['Icon', 'iconSide', 'onlyIcon'] },
+  },
+}
+
 export const ButtonWithIconAndLabel: StoryObj<IButtonProps> = {
   args: {
     iconSide: 'left',