+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @conquerpage/react
 
+## 2.0.1
+
+### Patch Changes
+
+- added a new button variant 'orange'
+
 ## 2.0.0
 
 ### Major Changes